Move Change struct to josh-changes
The struct is only used in that crate.

Change: mv-change
