+++ josh-changes/src/lib.rs
@@ -1,5 +1,41 @@
 use anyhow::anyhow;
-use josh_core::Change;
+
+#[derive(Debug)]
+struct Change {
+    author: String,
+    id: Option<String>,
+    requires: Vec<String>,
+    commit: git2::Oid,
+}
+
+impl Change {
+    fn new(commit: git2::Oid) -> Self {
+        Self {
+            author: Default::default(),
+            id: Default::default(),
+            requires: Default::default(),
+            commit,
+        }
+    }
+}
+
+fn get_change_id(commit: &git2::Commit) -> Change {
+    let mut change = Change::new(commit.id());
+    change.author = commit.author().email().unwrap_or("").to_string();
+
+    for line in commit.message().unwrap_or("").lines() {
+        if let Some(id) = line.strip_prefix("Change: ") {
+            change.id = Some(id.to_string());
+        }
+        if let Some(id) = line.strip_prefix("Change-Id: ") {
+            change.id = Some(id.to_string());
+        }
+        if let Some(id) = line.strip_prefix("Requires: ") {
+            change.requires.push(id.to_string());
+        }
+    }
+    change
+}
 
 #[derive(PartialEq, Clone, Debug)]
 pub enum PushMode {
@@ -124,14 +160,10 @@ pub fn downstack(
 
     // Parse Requires: footers, keeping only those referencing changes
     // actually present in the intermediates
-    let required_raw: std::collections::HashSet<String> = josh_core::get_change_id(&change_commit)
-        .requires
-        .into_iter()
-        .collect();
-    let intermediate_ids: Vec<Option<String>> = commits
-        .iter()
-        .map(|c| josh_core::get_change_id(c).id)
-        .collect();
+    let required_raw: std::collections::HashSet<String> =
+        get_change_id(&change_commit).requires.into_iter().collect();
+    let intermediate_ids: Vec<Option<String>> =
+        commits.iter().map(|c| get_change_id(c).id).collect();
     let available_ids: std::collections::HashSet<&str> = intermediate_ids
         .iter()
         .filter_map(|id| id.as_deref())
@@ -199,7 +231,7 @@ pub fn downstack(
     )
 }
 
-pub fn changes_to_refs(
+fn changes_to_refs(
     baseref: &str,
     change_author: &str,
     changes: Vec<Change>,
@@ -264,7 +296,7 @@ fn get_changes(
     let mut changes = vec![];
     for rev in walk {
         let commit = repo.find_commit(rev?)?;
-        changes.push(josh_core::get_change_id(&commit));
+        changes.push(get_change_id(&commit));
     }
 
     Ok(changes)
