+++ josh-core/src/lib.rs
@@ -25,25 +25,6 @@ pub mod housekeeping;
 pub mod link;
 pub mod submodules;
 
-#[derive(Debug)]
-pub struct Change {
-    pub author: String,
-    pub id: Option<String>,
-    pub requires: Vec<String>,
-    pub commit: git2::Oid,
-}
-
-impl Change {
-    fn new(commit: git2::Oid) -> Self {
-        Self {
-            author: Default::default(),
-            id: Default::default(),
-            requires: Default::default(),
-            commit,
-        }
-    }
-}
-
 #[derive(
     Clone, Hash, PartialEq, Eq, Copy, PartialOrd, Ord, Debug, serde::Serialize, serde::Deserialize,
 )]
@@ -160,27 +141,6 @@ impl $name {
     }
 }
     }
-}
-
-pub fn get_change_id(commit: &git2::Commit) -> Change {
-    let mut change = Change::new(commit.id());
-    change.author = commit.author().email().unwrap_or("").to_string();
-
-    let mut have_change_id = false;
-    for line in commit.message().unwrap_or("").split('\n') {
-        if !have_change_id && line.starts_with("Change: ") {
-            change.id = Some(line.replacen("Change: ", "", 1));
-            // If there is a "Change-Id" as well, it will take precedence
-        }
-        if !have_change_id && line.starts_with("Change-Id: ") {
-            change.id = Some(line.replacen("Change-Id: ", "", 1));
-            have_change_id = true;
-        }
-        if let Some(id) = line.strip_prefix("Requires: ") {
-            change.requires.push(id.to_string());
-        }
-    }
-    change
 }
 
 #[tracing::instrument(level = tracing::Level::TRACE, skip(transaction))]
